src/backend/requirements.txt
flask
openai
PyPDF2
xmltodict
python-dotenv